agent/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-dev-agent"
version = "0.1.0"
description = "Autonomous AI Development Agent — Jira/Git ticket to PR"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "PyGithub>=2.4",
    "python-gitlab>=4.5",
    "anthropic>=0.39",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
